Regroup and fix wordings
